requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
dataclasses==0.6
docopt==0.6.2
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.19.5
opencv-python==4.5.3.56
Pillow==8.3.1
pipreqs==0.4.10
pyzbar==0.1.8
qrcode==7.2
requests==2.26.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
yarg==0.1.9
zipp==3.5.0
